refactor: massive refactor of package list, removing the middleware class
