+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/Create Shipping Label Form/Package Details/Package Selection/ShippingLabelPackageListViewModel.swift
@@ -5,18 +5,15 @@ import Yosemite
 ///
 final class ShippingLabelPackageListViewModel: ObservableObject {
 
-    private(set) var state: ShippingLabelPackageSelectedViewModel.State
     private(set) var dimensionUnit: String
     private(set) var customPackages: [ShippingLabelCustomPackage]
     private(set) var predefinedOptions: [ShippingLabelPredefinedOption]
     @Published private(set) var selectedCustomPackage: ShippingLabelCustomPackage?
     @Published private(set) var selectedPredefinedPackage: ShippingLabelPredefinedPackage?
 
-    init(state: ShippingLabelPackageSelectedViewModel.State,
-         packagesResponse: ShippingLabelPackagesResponse?,
+    init(packagesResponse: ShippingLabelPackagesResponse?,
          selectedCustomPackage: ShippingLabelCustomPackage? = nil,
          selectedPredefinedPackage: ShippingLabelPredefinedPackage? = nil) {
-        self.state = state
         self.dimensionUnit = packagesResponse?.storeOptions.dimensionUnit ?? ""
         self.customPackages = packagesResponse?.customPackages ?? []
         self.predefinedOptions = packagesResponse?.predefinedOptions ?? []
@@ -26,8 +23,8 @@ final class ShippingLabelPackageListViewModel: ObservableObject {
 
     func didSelectCustomPackage(_ customPackage: ShippingLabelCustomPackage?) {
         guard customPackage != nil else { return }
-        selectedPredefinedPackage = nil
         selectedCustomPackage = customPackage
+        selectedPredefinedPackage = nil
     }
 
     func didSelectPredefinedPackage(_ predefinedPackage: ShippingLabelPredefinedPackage?) {
